fix(schematic): stop silently discarding coordinates and CDF meaning
Two defects from the 2026-09-10 tooling audit, both of the same shape: the
tool had the information and threw it away without saying so.

Defect B — coordinates were truncated to 0 and the call answered ok.
`json_i64_or(params.get("x"), 0)` gave `as_i64().unwrap_or(0)`, so both
`{"x": -1.5}` and `{"x": "7"}` placed the instance at the origin with
`status: ok`. The schematic grid is 0.0625; most real placements are not
integers. Coordinates are now f64 through dispatcher, ops, commands, editor,
CLI and batch spec — `dbCreateInst`'s `l_point` is in user units (IC23.1
skdfref.fnd) — and a present-but-non-numeric value is an error.

`move_instance` was already f64, so the two halves of "place it, then tidy
it" disagreed about what a coordinate is.

Same family, found while fixing it: `schematic.wire` parsed its points with
two `filter_map`s, so one mistyped coordinate silently shortened the wire
and still reported ok. Every point must now parse.

Live on IC23.1: place x=-1.5 y=4.0625 reads back as -1.5 / 4.0625;
x="7" errors instead of placing at the origin.

list_cdf_params — names without meaning.
It returned {name, type, value}, so a caller facing vsin's 135 names could
not tell `va` from `vaDBm` from `acm`. That is the gap that produced
`ampl=5m` on a vsin and a day chasing a flat transient. The CDF already
carries the answer, measured rather than assumed:

  prompt      135/135 on vsin, and on the PDK (`w` -> "Total Width")
  units       voltage / frequency / lengthMetric
  defValue    present; `w` defaults 300n while the instance holds 20u
  choices     1/135 (filenums), emitted as a JSON array when present
  description 0/135 — so prose still comes from analoglib.info

This is the only route that covers PDK devices, which no Cadence manual
documents. `simM = iPar("m")` survives the JSON escaping intact.

cargo test --release: 3203 passed / 0 failed / 7 ignored (was 3188/0/7).

(cherry picked from commit bc520d3e53460ee6c7452f1aa3bbb4a8d136a615)

+++ src/client/editor.rs
@@ -82,7 +82,7 @@ impl<'a> SchematicEditor<'a> {
         cell: &str,
         view: &str,
         name: &str,
-        origin: (i64, i64),
+        origin: (f64, f64),
         orient: &str,
     ) {
         let ops = SchematicOps;
@@ -90,18 +90,18 @@ impl<'a> SchematicEditor<'a> {
             .push(ops.create_instance(lib, cell, view, name, origin, orient));
     }
 
-    pub fn add_wire(&mut self, points: Vec<(i64, i64)>, layer: &str, net_name: &str) {
+    pub fn add_wire(&mut self, points: Vec<(f64, f64)>, layer: &str, net_name: &str) {
         let ops = SchematicOps;
         self.commands
             .push(ops.create_wire(&points, layer, net_name));
     }
 
-    pub fn add_label(&mut self, net_name: &str, origin: (i64, i64)) {
+    pub fn add_label(&mut self, net_name: &str, origin: (f64, f64)) {
         let ops = SchematicOps;
         self.commands.push(ops.create_wire_label(net_name, origin));
     }
 
-    pub fn add_pin(&mut self, net_name: &str, pin_type: &str, origin: (i64, i64)) {
+    pub fn add_pin(&mut self, net_name: &str, pin_type: &str, origin: (f64, f64)) {
         let ops = SchematicOps;
         self.commands
             .push(ops.create_pin(net_name, pin_type, origin));
